+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 4.3.3
 
+- feature: set verified system version to 4.3.9
 - feature: update fast-forward check based on the system's implementation in `applyKeybindings`
 - feature: improve DAE integration with an AR Messages tab in the field browser and forcing all message flags to use Custom change mode
 - feature: refactor messages to use flags on the actor instead of reading the active effects